images/spz/requirements.txt
alabaster==0.7.7
amqp==1.4.9
anyjson==0.3.3
argon2==0.1.10
Babel==2.2.0
billiard==3.3.0.22
blinker==1.4
celery==3.1.20
decorator==4.0.9
dnspython3==1.12.0
docutils==0.12
email_validator==1.0.1
flake8==2.5.2
flake8-coding==1.1.1
Flask==0.10.1
Flask-Assets==0.11
Flask-Cache==0.13.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
-e git+https://github.com/reingart/pyfpdf.git@3114b4f249bbaca199af6312f1cc982f8585ade9#egg=fpdf
idna==2.0
ipython==4.1.2
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
kombu==3.0.33
MarkupSafe==0.23
mccabe==0.4.0
path.py==8.1.2
pep8==1.7.0
pexpect==4.0.1
phonenumbers==7.2.6
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
pyflakes==1.1.0
Pygments==2.1.3
pyinotify==0.9.6
python-magic==0.4.10
pytz==2015.7
rcssmin==1.0.6
redis==2.10.5
requests==2.9.1
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.6
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.12
traitlets==4.1.0
uWSGI==2.0.12
webassets==0.10.1
Werkzeug==0.11.4
WTForms==2.1
